+++ src/components/utils/Json/Json.tsx
@@ -1,0 +1,5 @@
+import classes from "./Json.module.css";
+
+export default function Json<T>(obj: T) {
+    return <pre className={classes.pre}>{JSON.stringify(obj)}</pre>;
+}