+++ src/core/model/api-client.ts
@@ -1,0 +1,6 @@
+export type QueryClientProps = {
+  queryKeys: Array<string>;
+  options?: RequestInit;
+  path?: Array<string>;
+  queryParams?: Record<string, string | number | boolean>;
+};