+++ src/main.tsx
@@ -3,9 +3,13 @@ import { createRoot } from "react-dom/client";
 import "@/index.css";
 import { RouterProvider } from "@tanstack/react-router";
 import { router } from "@/core/router";
+import { QueryClientProvider } from "@tanstack/react-query";
+import { queryClient } from "@/core/query-client";
 
 createRoot(document.getElementById("root")!).render(
     <StrictMode>
-        <RouterProvider router={router} />
+        <QueryClientProvider client={queryClient}>
+            <RouterProvider router={router} />
+        </QueryClientProvider>
     </StrictMode>,
 );