+++ vite.config.ts
@@ -1,13 +1,21 @@
-import { defineConfig } from "vite";
+import { defineConfig } from "vitest/config";
 import react from "@vitejs/plugin-react";
 import path from "path";
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  envDir: "./env",
+  test: {
+    environment: "jsdom",
+    setupFiles: ["./src/test/setup.ts"],
+    isolate: true,
+    pool: "forks",
+  },
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
+      "@test": path.resolve(__dirname, "./test"),
     },
   },
 });