+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+## [0.2.0] - 2025-11-04
+
+### 🚀 Features
+
+- Configured tanstack-query
+- Fetch data util function
+- Api client hook
+- Centralised layout for fetched contents
+- Json component as debugging util
+- Layout with data for list and detail
+- Implemente query params in api-client
+
+### 🐛 Bug Fixes
+
+- Changelog generation with prepend
+- Changelog prepend path
+- Working prepend version on changelog
+
+### 📚 Documentation
+
+- Updatet changelog
+- Documented utils, apiclient, project structure and table of contents
+
+### 🧪 Testing
+
+- Configured for testing hooks
+- Configured centralized mocks
+
+### ⚙️ Miscellaneous Tasks
+
+- Project scaffolding
+- *(vite-config)* Removed tanstack router plugin
+- Removed unused package
+- Configured environment
+- Package.json syncronization on release
+- Added testing-library
+- Added jsom dependency
+- Configured alias for tests folder
 ## [0.1.0] - 2025-11-04
 
 ### ⚙️ Miscellaneous Tasks